% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_axisLonLat.R
\name{axisLonLat}
\alias{axisLonLat}
\title{Add Latitude and Longitude Axes to a UTM Map}
\usage{
axisLonLat(
  map.crs,
  xlim = NULL,
  ylim = NULL,
  by = 0.5,
  side = NULL,
  line = 0,
  tcl = -0.5
)
}
\arguments{
\item{map.crs}{A character string specifying the Proj4 code or an object of
class \code{CRS} of the plotted  map projection .}

\item{xlim}{A numeric vector of length 2 specifying the limits of the X-axis.
If \code{NULL} (the default), the function uses the current plot limits.}

\item{ylim}{A numeric vector of length 2 specifying the limits of the Y-axis.
If \code{NULL} (the default), the function uses the current plot limits.}

\item{by}{A numeric value specifying the interval for the axis labels, in
degrees.}

\item{side}{An integer (1, 2, 3, or 4) indicating which side of the plot the
axis should be drawn on. If \code{NULL} (the default), the function will draw
axes on both the X and Y sides of the plot.}

\item{line}{A numeric value specifying the line position of the axis labels.
Default is 0, which places the labels at the default position. Positive
values move the labels outward, and negative values move them inward.}

\item{tcl}{A numeric value specifying the tick length.}
}
\value{
This function adds latitude and longitude axis labels to the
current plot. It does not return any value.
}
\description{
This function adds axes to a map, displaying coordinates in latitude and
longitude format (degrees and minutes), for a map projected in UTM.
}
