% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectFilter.R
\name{detectFilter}
\alias{detectFilter}
\title{Identify false or spurious detections in acoustic telemetry data}
\usage{
detectFilter(
  tag.id,
  time.stamp,
  rec.id = NULL,
  time.int = 24,
  units = "hours",
  return_both = FALSE
)
}
\arguments{
\item{tag.id}{vector with the ID codes of tagged individuals.}

\item{time.stamp}{vector with the date and time of each detection in
\code{POSIXt} format.}

\item{rec.id}{vector with the ID codes of the receivers for each detection.
Optional argument, if provided, the fuction will also identify the false
detections within each receiver.}

\item{time.int}{time interval (in hours) that define false detections. It is
set to 24h by default, as it is the most commonly used value.}

\item{return_both}{A logical value indicating whether to return both the
detection results for the whole array and individual receivers (if
\code{rec.id} is provided). If \code{TRUE}, the function returns both
results. If \code{FALSE} (default), it returns only the results for the
entire array or for individual receivers.}
}
\value{
The function returns a data.frame with one or two logical vectors
indicating if each detection has been identified as good (\code{TRUE}) or
false (\code{FALSE}), taking into account the whole detection
\code{array} (first column) or each \code{receiver} separately (optional
second column, only if \code{receiver} is provided).
}
\description{
False detections are defined as single detections within a predefined time
interval (usually 24h). They can be identified among all the detections of
the whole acoustic array or individually (providing the \code{receiver}
argument), in which case the the detections of each single acoustic receiver
are considered independently.
}
\examples{

# Create a random dataset
set.seed(900)
time_stamp <- Sys.time() + c(1:24, 70:71, 200, 300:320, 500, 701:720) * 3600
tag_id <- factor(sample(c("i1", "i2"), length(time.stamp), replace = TRUE))
rec_id <- factor(sample(c("r1", "r2"), length(time.stamp), replace = TRUE))

# False detections for the whole receiver array
f1 <- detectFilter(tag_id, time_stamp)
plot(time_stamp, tag_id, pch = (1:2)[rec.id])
points(time_stamp[!f1], tag_id[!f1], col = 2, cex = 2)

# False detections within individual receivers
f2 <- detectFilter(tag.id, time.stamp, rec.id, return_both = TRUE)
plot(time_stamp, tag_id, pch = (1:2)[rec.id])
points(time_stamp[!f2$receiver], tag_id[!f2$receiver], col = 4, cex = 2.4)
points(time_stamp[!f2$array], tag_id[!f2$array], col = 2, cex = 2)

}
