% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectFilter.R
\name{filtSubset}
\alias{filtSubset}
\title{Identify Valid Detections in a Subset of Data}
\usage{
filtSubset(time.stamp, time.int = 24, units = "hours")
}
\arguments{
\item{time.stamp}{A vector of timestamps (\code{POSIXt} format) representing the
detections of a single transmitter or subset for an individual receiver.}

\item{time.int}{A numeric value specifying the time interval threshold.
Defaults to 24.}

\item{units}{A character string specifying the time units for \code{time.int}.
Defaults to \code{"hours"}.}
}
\value{
A logical vector of the same length as \code{time.stamp}, where \code{TRUE}
indicates a valid detection and \code{FALSE} indicates an isolated (potentially
false) detection.
}
\description{
This helper function identifies valid detections in a subset of acoustic
telemetry data. A valid detection is one that is not isolated beyond a given
time interval (\code{time.int}) from its neighbors. This function is used within
\code{detectFilter()}.
}
