% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainVal.R
\name{mainVal}
\alias{mainVal}
\title{Select the Most Abundant Value in a Vector}
\usage{
mainVal(values)
}
\arguments{
\item{values}{A vector of values, usually station names or receiver IDs.}
}
\value{
The most abundant value from the input vector.
}
\description{
This function identifies and returns the most abundant value in a vector.
It calculates the frequency of each value and selects the one with the
maximum count. If there is a tie (i.e., multiple values have the same
maximum count), one of the tied values is selected randomly. This function
can be useful for selecting the receiver or station with the largest number
of detections within a time interval.
}
\examples{
station_ids <- c("st1", "st2", "st2", "st3", "st3")
mainVal(station_ids)

}
