% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milliSec.R
\name{milliSec}
\alias{milliSec}
\title{Function to extract the milliseconds from time stamps}
\usage{
milliSec(time.stamps)
}
\arguments{
\item{time.stamps}{Detection time stamps, POSIXct.}
}
\value{
A numeric vector of milliseconds (0 to 1 range).
}
\description{
This function extracts the millisecond componente of detection time stamps.
}
