% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale.R
\name{rescale}
\alias{rescale}
\title{Rescale variables}
\usage{
rescale(values, new.range, orig.range = NULL)
}
\arguments{
\item{values}{A numeric vector of values to be rescaled.}

\item{new.range}{A numeric vector of length 2 specifying the minimum and
maximum values of the new range.}

\item{orig.range}{A numeric vector of length 2 specifying the original range
of the values. If NULL (default), the function uses the
range of \code{values}.}
}
\value{
A numeric vector with rescaled values.
}
\description{
This function rescales a numeric vector to a specified new range. The
rescaling is based on either the range of the input values or a
user-provided original range.
}
