% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundSpeed.R
\name{soundSpeed}
\alias{soundSpeed}
\title{Function to calculate the speed of sound in water applying the Mackenzie
equation}
\usage{
soundSpeed(temp = 21, sal = 38, depth = 20)
}
\arguments{
\item{temp}{Numeric vector or single value of temperature, in celsius degrees.}

\item{sal}{Numeric vector or single value of salinity, in PSU.}

\item{depth}{Numeric vector or single value of depth.}
}
\value{
A numeric vector of sound speed in water, in m/s.
}
\description{
This function to estimates sound speed by applying a simplified version of
the Mackenzie equation (1981), given the temperature (temp), salinity (sal) and
depth.
Reference: Mackenzie, K.V. (1981). Nine-term equation for the sound speed in
the oceans. J. Acoust. Soc. Am. 70(3): 807-812.
https://doi.org/10.1121/1.386920
}
