% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFilter.R
\name{speedCalc}
\alias{speedCalc}
\title{Calculate speed from movement trajectory}
\usage{
speedCalc(x, y, time.stamps)
}
\arguments{
\item{x}{A numeric vector of x-coordinates (UTM eastings, in meters).}

\item{y}{A numeric vector of y-coordinates (UTM northings, in meters).}

\item{time.stamps}{A POSIXct vector of timestamps corresponding to each
position.}
}
\value{
A numeric vector of speeds (m/s), with NA as the first value.
}
\description{
This function calculates the speed (in meters per second) between
consecutive positions in a movement trajectory using UTM coordinates.
}
\examples{
x <- c(500000, 500010, 500020)
y <- c(4500000, 4500005, 4500015)
date_time <- as.POSIXct(c("2025-01-01 12:00:00", "2025-01-01 12:00:10",
                          "2025-01-01 12:00:20"))
speedCalc(x, y, date_time)

}
