% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timechunks.R
\name{timeChunks}
\alias{timeChunks}
\title{Divide time stamps into time chunks}
\usage{
timeChunks(time.stamps, time.thresh = 60, units = "mins")
}
\arguments{
\item{time.stamps}{A vector of time stamps in \code{POSIXct} format.}

\item{units}{A character string specifying the time units for \code{time.thresh}.
Defaults to \code{"mins"}.}

\item{time.thres}{A numeric value specifying the time interval used to define
gaps between chunks. If the time difference between consecutive detections
exceeds this threshold, a new chunk is started.}
}
\value{
A numeric vector assigning each time stamp to a chunk. The sequence
starts at 1 and increments whenever a gap larger than \code{time.thresh} is
found.
}
\description{
This function identifies contiguous chunks of time based on gaps between
consecutive time stamps. A new chunk is created whenever the time gap between
time stamps exceeds a specified threshold.
}
\examples{
date_time <- as.POSIXct(c("2023-01-01 00:00:00", "2023-01-01 00:05:00",
                          "2023-01-01 00:20:00", "2023-01-01 03:00:00",
                          "2023-01-01 03:45:00", "2023-01-01 07:00:00"))
timeChunks(date_time, time.thresh = 60, units = "mins")


}
