% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trackFilter.R
\name{turnAng}
\alias{turnAng}
\title{Calculate turning angles}
\usage{
turnAng(x, y, rad = FALSE)
}
\arguments{
\item{x}{A numeric vector of x-coordinates (UTM eastings, in meters).}

\item{y}{A numeric vector of y-coordinates (UTM northings, in meters).}

\item{rad}{Logical. If TRUE, returns the turning angle in radians; otherwise,
in degrees. Defaults to FALSE (degrees).}
}
\value{
A numeric vector of turning angles, with NA for the first and last
values.
}
\description{
This function computes the turning angle (change in movement direction)
between consecutive steps
}
\details{
This function calculates the turning angle between consecutive movement
steps in a trajectory using UTM coordinates.The turning angle is defined as
the difference between the bearing of the previous step (from the preceding
position to the current position) and the bearing of the following step (from
the current position to the next position). A positive angle indicates a
right turn, while a negative angle indicates a left turn. The first and last
positions will return \code{NA} since they do not have both a preceding and a
following step.
}
\examples{
x <- c(500000, 500010, 500020, 500015, 500020)
y <- c(4500000, 4500005, 4500015, 4500020, 4500025)
turnAng(x, y)

}
