% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrapDrift.R
\name{unwrapDrift}
\alias{unwrapDrift}
\title{Function to unwrap time drift}
\usage{
unwrapDrift(ms, time.stamps, granularity = 1)
}
\arguments{
\item{ms}{Vector of millisecond values.}

\item{time.stamps}{Detection time stamps, POSIXct.}

\item{granularity}{Signal emission granularity.}
}
\value{
A numeric vector of a continuous time drift.
}
\description{
Function to unwrap the drift in milliseconds (ranging between 0 and 1) to
extract the continuous drift pattern. It takes into account the granularity
of the signal emission period, that is, the precision of the time-intervals
at which signals are emitted. Thelma Biotel's transmitters have a granularity
of 1 (signals are always emitted at entire seconds in their internal clock),
while Lotek transmitters have a granularity of 0.5 (signals are emitted at
possible intervals of 0.5 seconds).
}
