% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_axisLonLat.R
\name{axisLonLat}
\alias{axisLonLat}
\title{Plot axes in Latitude and Longitude format}
\usage{
axisLonLat(
  map.crs,
  xlim = NULL,
  ylim = NULL,
  by = 0.5,
  side = NULL,
  line = 0,
  tcl = -0.5
)
}
\arguments{
\item{map.crs}{A character string specifying the Proj4 code of the original
map projection or an object of class \code{CRS}.}

\item{xlim}{A numeric vector of length 2 specifying the limits of the X-axis
(optional). Defaults to the current plot limits.}

\item{ylim}{A numeric vector of length 2 specifying the limits of the Y-axis
(optional). Defaults to the current plot limits.}

\item{by}{A numeric value specifying the interval (in degrees) for the axis
labels. Defaults to 0.5 degrees.}

\item{side}{An integer (1, 2, 3, or 4) indicating which side of the plot the
axis should be drawn on. Defaults to NULL, which draws both X
and Y axes.}

\item{line}{A numeric value specifying the line position of the axis labels.
Defaults to 0.}

\item{tcl}{A numeric value specifying the tick length (negative values point
inward). Defaults to -0.5.}
}
\value{
This function adds latitude and longitude axis labels to the current plot.
}
\description{
This function draws the axes of a UTM map in Latitude and Longitude format.
}
