% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectFilter.R
\name{detectFilter}
\alias{detectFilter}
\title{Identify Valid and False or Spurious Detections in Acoustic Telemetry Data}
\usage{
detectFilter(
  tag.id,
  time.stamp,
  rec.id = NULL,
  time.int = 24,
  units = "hours",
  return.both = FALSE
)
}
\arguments{
\item{tag.id}{A vector containing the ID codes of tagged individuals or
transmitters.}

\item{time.stamp}{A vector of detection time stamps in \code{POSIXt} format.}

\item{rec.id}{(Optional) A vector containing the receiver IDs for each
detection. If provided, false detections will be identified within each
receiver separately.}

\item{time.int}{A numeric value specifying the time interval used to define
false detections.}

\item{units}{A character string specifying the time units for \code{time.int.}
Defaults to \code{"hours"}.}

\item{return.both}{Logical. If \code{TRUE}, the function returns a data frame
containing false detection indicators for both the entire receiver
array and individual receivers (if \code{rec.id} is provided). If
\code{FALSE} (default), only a single logical vector is returned.}
}
\value{
A logical vector or data frame indicating whether each detection is
valid (\code{TRUE}) or false (\code{FALSE}).
- If \code{rec.id} is not provided, the function returns a single logical
vector corresponding to the entire receiver array.
- If \code{rec.id} is provided and \code{return_both = FALSE}, the function returns
only the logical vector for individual receivers.
- If \code{rec.id} is provided and \code{return_both = TRUE}, the function returns
a data frame with two logical vectors: one for the entire array and one
for individual receivers.
}
\description{
This function identifies valid and false detections in acoustic telemetry
data based on a predefined time interval (default: 24 hours). A false
detection is defined as an isolated detection occurring within the specified
time interval. The function can analyze detections across all receivers or
separately for each receiver deployment (if the \code{rec.id} argument is
provided).
}
\examples{
# Generate a random dataset
set.seed(900)
time_stamp <- Sys.time() + c(1:24, 70:71, 200, 300:320, 500, 701:720) * 3600
tag_id <- factor(sample(c("i1", "i2"), length(time_stamp), replace = TRUE))
rec_id <- factor(sample(c("r1", "r2"), length(time_stamp), replace = TRUE))

# Identify false detections for the entire receiver array
f1 <- detectFilter(tag_id, time_stamp)
plot(time_stamp, tag_id, pch = (1:2)[rec_id])
points(time_stamp[!f1], tag_id[!f1], col = 2, cex = 2)

# Identify false detections within individual receivers
f2 <- detectFilter(tag_id, time_stamp, rec_id, time.int = 1, units = "days",
                   return.both = TRUE)
plot(time_stamp, tag_id, pch = (1:2)[rec_id])
points(time_stamp[!f2$receiver], tag_id[!f2$receiver], col = 4, cex = 2.4)
points(time_stamp[!f2$array], tag_id[!f2$array], col = 2, cex = 2)

}
