% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unwrapDrift.R
\name{unwrapDrift}
\alias{unwrapDrift}
\title{Unwrap Time Drift in Acoustic Telemetry Data}
\usage{
unwrapDrift(ms, time.stamps, granularity = 1)
}
\arguments{
\item{ms}{A numeric vector of millisecond values, ranging from 0 to 1.}

\item{time.stamps}{A \code{POSIXct} vector of detection time stamps.}

\item{granularity}{A numeric value specifying the signal emission
granularity. Defaults to \code{1} for Thelma Biotel transmitters. Use \code{0.5} for
Lotek transmitters.}
}
\value{
A numeric vector representing the unwrapped, continuous time drift.
}
\description{
This function reconstructs the continuous time drift pattern from a vector
of millisecond values that range between 0 and 1. Since the observed drift
is cyclical (0-1 range), the function unwraps it by correcting for
discontinuities.The correction accounts for the signal emission granularity
of the transmitter, which refers to the precision of the signal emission
intervals. The function estimates the mean slope of the drift using time
differences between consecutive detections. It only considers pairs of
detections with small drift changes (less than half the granularity) and
those occurring within one hour of each other.
}
